refactor: 유저 페이지 폴더 구조 개선
